+++ app/javascript/application_webpack.js
@@ -49,9 +49,14 @@ Chart.register(...registerables);
 window.Chart = Chart;
 import "javascripts/chart_config";
 
+// flatpickr
+import flatpickr from "flatpickr";
+window.flatpickr = flatpickr;
+
 window.Routes = require("./routes");
 
 // assets with side-effects only
+import "javascripts/flatpickr_config";
 import "javascripts/help-system";
 import "javascripts/layouts";
 import "javascripts/menu";